fix: confine PCL to cloud_mesher instead of the whole sensors package
The colcon build of sensors broke, and it took topic_remapper down with it --
a target that has nothing to do with meshing. Two separate leaks, both from
treating PCL as a package-wide dependency when it belongs to one executable.

pcl_conversions went into THIS_PACKAGE_INCLUDE_DEPENDS, which
ament_target_dependencies applies to every target in the package. Its ament
export carries the whole of PCL, VTK included, so topic_remapper started linking
against libnl for rendering code it never calls.

find_package(PCL) reaches further than it looks: it calls link_directories(),
which applies to every target defined after it in the same scope, and it reports
/usr/include among its include directories. That second one is the nastier of the
two -- this toolchain ships its own glibc, and mixing the host's headers with it
makes pthread_cond_t resolve to the wrong type, so <mutex> fails to compile with
an error pointing into libstdc++ that says nothing about the cause.

Both are now in src/cloud_mesher/CMakeLists.txt, where a directory scope contains
them: PCL, pcl_conversions and visualization_msgs are found there, /usr/include
is declared implicit there, and the package-wide dependency list goes back to
what it was. The one target that wants PCL is the only one that sees it.

Also linked against the five PCL component libraries actually used rather than
the aggregate ${PCL_LIBRARIES}, with --allow-shlib-undefined so VTK's own
undefined symbols are left to the loader, and fixed a real -Wsign-compare in the
vertex bounds check (polygon indices are uint32_t against a size_t).

Verified both ways from clean: colcon builds sensors with all four executables,
nix builds it, and the node still produces 21567 triangles from a 20451-point map
in 0.23 s with /mesh coming out as type 11.

+++ software/ros_ws/src/sensors/src/cloud_mesher/cloud_mesher.cpp
@@ -248,7 +248,7 @@ private:
       }
       bool in_range = true;
       for (const auto idx : polygon.vertices) {
-        in_range = in_range && idx < n_vertices;
+        in_range = in_range && static_cast<size_t>(idx) < n_vertices;
       }
       if (!in_range) {
         continue;
